requirements.txt
torch
torchvision
numpy
matplotlib
seaborn
tqdm
scikit-learn
pyyaml
